arrumando a funcao de deletar imagem, implementacao inicial de api key para acesso, e adicionado as variaveis no arquivo env

+++ deleteimage.js
@@ -1,19 +1,24 @@
 const fs = require("fs");
 
 const deleteImage = (imagem) => {
-  let img = imagem;
 
-  if (!img.includes("uploads/")) {
-    img = "uploads/" + img;
-  }
-
-  fs.unlink(img, err => {
-    if (err) {
-      console.error("Erro ao deletar a imagem");
-      return;
+  if (imagem) {
+    let img = imagem;
+    
+    if (!img.includes("uploads\\")) {
+      img = "uploads\\" + img;
     }
-    console.log("Imagem deletada com sucesso.");
-  });
+  
+    fs.unlink(img, err => {
+      if (err) {
+        console.error("Erro ao deletar a imagem");
+        return;
+      }
+      console.log("Imagem deletada com sucesso.");
+    });
+  } else {
+    console.log("Imagem inexistente");
+  }
 }
 
 module.exports = {
